requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
firebase-admin==6.2.0
channels==4.0.0
channels-redis==4.1.0
stripe==7.6.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.5.0
django-storages==1.14.2
boto3==1.28.64
dj-database-url==2.1.0 